CODEBASE: Move Availability and Attempt to a more generic place
