+++ src/Bladeburner/Actions/Action.ts
@@ -1,6 +1,7 @@
 import type { Bladeburner } from "../Bladeburner";
 import type { Person } from "../../PersonObjects/Person";
-import type { Availability, SuccessChanceParams } from "../Types";
+import type { SuccessChanceParams } from "../Types";
+import type { Availability } from "../../types";
 import type { Skills as PersonSkills } from "../../PersonObjects/Skills";
 
 import { addOffset } from "../../utils/helpers/addOffset";
