+++ src/Bladeburner/Actions/BlackOperation.ts
@@ -1,5 +1,6 @@
 import type { Bladeburner } from "../Bladeburner";
-import type { ActionIdFor, Availability } from "../Types";
+import type { ActionIdFor } from "../Types";
+import type { Availability } from "../../types";
 
 import { BladeburnerActionType, BladeburnerBlackOpName } from "@enums";
 import { ActionClass, ActionParams } from "./Action";
