+++ src/Bladeburner/Actions/LevelableAction.ts
@@ -1,6 +1,6 @@
 import type { Bladeburner } from "../Bladeburner";
 import type { IReviverValue } from "../../utils/JSONReviver";
-import type { Availability } from "../Types";
+import type { Availability } from "../../types";
 
 import { ActionClass, ActionParams } from "./Action";
 import { getRandomIntInclusive } from "../../utils/helpers/getRandomIntInclusive";
